refactor: move sanitize to output module and promote --stage to error
- Move sanitize_for_terminal from repl::sanitize to output::sanitize so it
  is not coupled to the REPL when used by main.rs, explain, and future modes
- Add src/output/mod.rs as home for shared terminal output utilities
- Update all import paths in repl/mod.rs, repl/custom/{history,layers}.rs,
  and main.rs
- Promote --stage from a silent warning to a fatal error: proceeding without
  stage selection would show the wrong stage without any indication to the user
- Update --stage integration test to assert exit 1 and the error message
