+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod engine;
 pub mod explain;
 pub mod model;
+pub mod output;
 pub mod parser;
 pub mod repl;
